+++ packages/adapter-interfaces/src/adapters/RefreshTokens.ts
@@ -1,16 +1,19 @@
 import { RefreshToken, RefreshTokenInsert, Totals, ListParams } from "../types";
 
-interface RefreshTokenResponse extends Totals {
+export interface ListRefreshTokenResponse extends Totals {
   refresh_tokens: RefreshToken[];
 }
 
 export interface RefreshTokensAdapter {
   create: (
     tenant_id: string,
-    session: RefreshTokenInsert,
+    refresh_token: RefreshTokenInsert,
   ) => Promise<RefreshToken>;
   get: (tenant_id: string, id: string) => Promise<RefreshToken | null>;
-  list(tenantId: string, params?: ListParams): Promise<RefreshTokenResponse>;
+  list(
+    tenantId: string,
+    params?: ListParams,
+  ): Promise<ListRefreshTokenResponse>;
   update: (
     tenant_id: string,
     id: string,