+++ packages/kysely/src/refreshTokens/create.ts
@@ -1,0 +1,23 @@
+import { Kysely } from "kysely";
+import { RefreshToken, RefreshTokenInsert } from "@authhero/adapter-interfaces";
+import { Database } from "../db";
+
+export function create(db: Kysely<Database>) {
+  return async (
+    tenant_id: string,
+    refreshToken: RefreshTokenInsert,
+  ): Promise<RefreshToken> => {
+    const createdRefreshToken = {
+      ...refreshToken,
+      created_at: new Date().toISOString(),
+      expires_at: new Date().toISOString(),
+    };
+
+    await db
+      .insertInto("refresh_tokens")
+      .values({ ...createdRefreshToken, tenant_id })
+      .execute();
+
+    return { ...refreshToken, ...createdRefreshToken };
+  };
+}