+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,27 @@
 # authhero
 
+## 0.51.0
+
+### Minor Changes
+
+- Add refresh tokens to jwt
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.40.0
+
+## 0.50.0
+
+### Minor Changes
+
+- Store refresh tokesn
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.39.0
+
 ## 0.49.1
 
 ### Patch Changes